third_party/WebKit/LayoutTests/media/autoplay-muted.html

<!DOCTYPE html>
<title>Test for autoplay of muted video</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="media-file.js"></script>
<script>
    window.internals.settings.setMediaPlaybackRequiresUserGesture(true);
    window.internals.runtimeFlags.autoplayMutedVideosEnabled = true;

    function createMutedVideoElement() {
      var e = document.createElement('video');
      e.src = findMediaFile('video', 'content/test');
      e.muted = true;
      return e;
    }

    async_test(function(t)
    {
        var e = createMutedVideoElement();
        e.autoplay = true;

        var expectedEvents = [
            'canplay', 'play', 'playing'];
        var eventWatcher = new EventWatcher(t, e, expectedEvents);
        eventWatcher.wait_for(expectedEvents)
            .then(t.step_func_done(function()
                  {
                      assert_false(e.paused);
                  }));
    }, "muted-autoplay");

    promise_test(function()
    {
        return createMutedVideoElement().play();
    }, "muted-playjs");
</script>
